db_redis: add string key length protection
Introduce a list of improvements for size overflow
protection:
- change func signatures to work with `size_t` only
- check the raw pointer and its length

This measure will protect inserts from extremely
large key size (e.g. possible negative length that
then gets into integer overflow) and hence from
triggering the stack canary.

+++ src/modules/db_redis/redis_dbase.c
@@ -914,7 +914,7 @@ static int db_redis_scan_query_keys_pattern(km_redis_con_t *con,
 			LM_ERR("Failed to print integer for scan query\n");
 			goto err;
 		}
-		if(db_redis_key_add_string(&query_v, match_count_str, l) != 0) {
+		if(db_redis_key_add_string(&query_v, match_count_str, (size_t)l) != 0) {
 			LM_ERR("Failed to add count value to scan query\n");
 			goto err;
 		}
